Organizes directory structure
